+++ tapo/src/responses/smart_cam.rs
@@ -6,6 +6,9 @@ mod rtsp_stream_url;
 mod snapshot_result;
 mod timezone_hub_result;
 
+#[cfg(feature = "debug")]
+mod app_component_list_result;
+
 pub use general_device_list_hub_result::*;
 pub use recording_download_result::*;
 pub use recordings_hub_result::*;
@@ -14,3 +17,6 @@ pub use snapshot_result::*;
 pub use timezone_hub_result::*;
 
 pub(crate) use control_child_result::*;
+
+#[cfg(feature = "debug")]
+pub(crate) use app_component_list_result::*;